+++ llm_guard/input_scanners/code.py
@@ -12,7 +12,9 @@ LOGGER = get_logger()
 
 DEFAULT_MODEL = Model(
     path="philomath-1209/programming-language-identification",
+    revision="9090d38e7333a2c6ff00f154ab981a549842c20f",
     onnx_path="philomath-1209/programming-language-identification",
+    onnx_revision="9090d38e7333a2c6ff00f154ab981a549842c20f",
     onnx_subfolder="onnx",
     pipeline_kwargs={"truncation": True},
 )