+++ llm_guard/input_scanners/language.py
@@ -11,7 +11,9 @@ LOGGER = get_logger()
 
 DEFAULT_MODEL = Model(
     path="papluca/xlm-roberta-base-language-detection",
+    revision="9865598389ca9d95637462f743f683b51d75b87b",
     onnx_path="ProtectAI/xlm-roberta-base-language-detection-onnx",
+    onnx_revision="dce2fa14a0dc61b6f889537e9ad4fccf083b22bd",
     pipeline_kwargs={
         "max_length": 512,
         "truncation": True,