+++ llm_guard/model.py
@@ -10,7 +10,9 @@ class Model:
     Attributes:
         path (str): Path to the model.
         subfolder (str): Subfolder in the model path.
+        revision (str, optional): Revision of the model.
         onnx_path (str, optional): Path to the ONNX model.
+        onnx_revision (str, optional): Revision of the ONNX model.
         onnx_subfolder (str): Subfolder in the ONNX model path.
         kwargs (Dict, optional): Keyword arguments passed to the model (transformers).
         pipeline_kwargs (Dict, optional): Keyword arguments passed to the pipeline (transformers).
@@ -18,7 +20,9 @@ class Model:
 
     path: str
     subfolder: str = ""
+    revision: Optional[str] = None
     onnx_path: Optional[str] = None
+    onnx_revision: Optional[str] = None
     onnx_subfolder: str = ""
     onnx_filename: str = "model.onnx"
     kwargs: Dict = dataclasses.field(default_factory=dict)